requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.0
opencv-python-headless>=4.9.0
Pillow>=10.3.0
numpy>=1.26.0
pandas>=2.2.0
openpyxl>=3.1.2
plotly>=5.22.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-mock>=3.14.0
pytest-html>=4.1.1
requests>=2.31.0
fpdf2>=2.7.9
matplotlib>=3.8.0
